refactor lv4
