--- PokemonContactApp/Modules/AddContact/Model/ContactModel.swift
@@ -1,6 +1,0 @@
-//
-//  ContactModel.swift
-//  PokemonContactApp
-//
-//  Created by 김이든 on 7/10/25.
-//
