+++ PokemonContactApp/Modules/AddContact/Model/ImageResponse.swift
@@ -8,10 +8,6 @@
 import Foundation
 
 struct ImageResponse: Decodable {
-    let id: Int
-    let name: String
-    let height: Int
-    let weight: Int
     let sprites: Sprites
 }
